Fixed cursor movement
