+++ Assets/Packages/com.studio23.ss2.bettercursormanager/Samples/Demo1/TextMesh Pro/Documentation.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 8e7e8f5a82a3a134e91c54efd2274ea9
+guid: 8527b74edb5404445b9f2dd567e97e78
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
