+++ Assets/Packages/com.studio23.ss2.bettercursormanager/Samples/Demo1/TextMesh Pro/Documentation/TextMesh Pro User Guide 2016.pdf.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 1b8d251f9af63b746bf2f7ffe00ebb9b
+guid: 5ae01a7a79cdd3a4192098411dd1daaa
 DefaultImporter:
   externalObjects: {}
   userData:
