+++ Assets/Packages/com.studio23.ss2.bettercursormanager/Samples/Demo1/TextMesh Pro/Fonts.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 6ab70aee4d56447429c680537fbf93ed
+guid: 682673cb957f51f40b374d3ce6979ecb
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
