+++ Assets/Packages/com.studio23.ss2.bettercursormanager/Samples/Demo1/TextMesh Pro/Fonts/LiberationSans - OFL.txt.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 6e59c59b81ab47f9b6ec5781fa725d2c
+guid: 0a4f09ce0bc4ead44bbf78d42043d188
 timeCreated: 1484171296
 licenseType: Pro
 TextScriptImporter:
