+++ Assets/Packages/com.studio23.ss2.bettercursormanager/Samples/Demo1/TextMesh Pro/Fonts/LiberationSans.ttf.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: e3265ab4bf004d28a9537516768c1c75
+guid: b96a024f6846bba4c9d74736fdb8304e
 timeCreated: 1484171297
 licenseType: Pro
 TrueTypeFontImporter:
