+++ Assets/Packages/com.studio23.ss2.bettercursormanager/Samples/Demo1/TextMesh Pro/Resources.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 243e06394e614e5d99fab26083b707fa
+guid: 5ebd9c0cf219e4f4c8fb627c409c6d2f
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
