+++ Assets/Packages/com.studio23.ss2.bettercursormanager/Samples/Demo1/TextMesh Pro/Resources/Fonts & Materials.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 731f1baa9d144a9897cb1d341c2092b8
+guid: 22c136e32874b9747ac73998789ecbe5
 folderAsset: yes
 timeCreated: 1442040525
 licenseType: Pro
