+++ Assets/Packages/com.studio23.ss2.bettercursormanager/Samples/Demo1/TextMesh Pro/Resources/Fonts & Materials/LiberationSans SDF - Drop Shadow.mat.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: e73a58f6e2794ae7b1b7e50b7fb811b0
+guid: ae296bf4c43a7804db917ed173f04184
 timeCreated: 1484172806
 licenseType: Pro
 NativeFormatImporter:
