+++ Assets/Packages/com.studio23.ss2.bettercursormanager/Samples/Demo1/TextMesh Pro/Resources/Fonts & Materials/LiberationSans SDF - Fallback.asset.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 2e498d1c8094910479dc3e1b768306a4
+guid: 0460c62aeb8a75f499826ee10698fa69
 timeCreated: 1484171803
 licenseType: Pro
 NativeFormatImporter:
