+++ Assets/Packages/com.studio23.ss2.bettercursormanager/Samples/Demo1/TextMesh Pro/Resources/Fonts & Materials/LiberationSans SDF - Outline.mat.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 79459efec17a4d00a321bdcc27bbc385
+guid: ee2f2d47fbad7cb4fa06233cb80b99aa
 timeCreated: 1484172856
 licenseType: Pro
 NativeFormatImporter:
