+++ Assets/Packages/com.studio23.ss2.bettercursormanager/Samples/Demo1/TextMesh Pro/Resources/Fonts & Materials/LiberationSans SDF.asset.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 8f586378b4e144a9851e7b34d9b748ee
+guid: 8780c5abfd19a4b41badec72555daa7e
 timeCreated: 1484171803
 licenseType: Pro
 NativeFormatImporter:
