+++ Assets/Packages/com.studio23.ss2.bettercursormanager/Samples/Demo1/TextMesh Pro/Resources/LineBreaking Following Characters.txt.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: fade42e8bc714b018fac513c043d323b
+guid: 3b6ffd812aa36ff498a71f1c0473602a
 timeCreated: 1425440388
 licenseType: Store
 TextScriptImporter:
