+++ Assets/Packages/com.studio23.ss2.bettercursormanager/Samples/Demo1/TextMesh Pro/Resources/LineBreaking Leading Characters.txt.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: d82c1b31c7e74239bff1220585707d2b
+guid: 332b37c61e8bdb34d9a4634376a17353
 timeCreated: 1425440388
 licenseType: Store
 TextScriptImporter:
