+++ Assets/Packages/com.studio23.ss2.bettercursormanager/Samples/Demo1/TextMesh Pro/Resources/Sprite Assets.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 512a49d95c0c4332bdd98131869c23c9
+guid: 2c89eb03805044b499060ed3bf9e5b07
 folderAsset: yes
 timeCreated: 1441876896
 licenseType: Pro
