+++ Assets/Packages/com.studio23.ss2.bettercursormanager/Samples/Demo1/TextMesh Pro/Resources/Sprite Assets/EmojiOne.asset.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: c41005c129ba4d66911b75229fd70b45
+guid: 7e4baa290079869489ca9612d886cf09
 timeCreated: 1480316912
 licenseType: Pro
 NativeFormatImporter:
