+++ Assets/Packages/com.studio23.ss2.bettercursormanager/Samples/Demo1/TextMesh Pro/Resources/Style Sheets.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 4aecb92fff08436c8303b10eab8da368
+guid: 767565fd2c2ec3442b4799656c47036f
 folderAsset: yes
 timeCreated: 1441876950
 licenseType: Pro
