+++ Assets/Packages/com.studio23.ss2.bettercursormanager/Samples/Demo1/TextMesh Pro/Resources/Style Sheets/Default Style Sheet.asset.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: f952c082cb03451daed3ee968ac6c63e
+guid: 76470d6adb53ac14bb9caf76f99cf1e9
 timeCreated: 1432805430
 licenseType: Store
 NativeFormatImporter:
