+++ Assets/Packages/com.studio23.ss2.bettercursormanager/Samples/Demo1/TextMesh Pro/Resources/TMP Settings.asset.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 3f5b5dff67a942289a9defa416b206f3
+guid: 5a0094c3a50dca946a852b7d0a354b80
 timeCreated: 1436653997
 licenseType: Pro
 NativeFormatImporter:
