+++ Assets/Packages/com.studio23.ss2.bettercursormanager/Samples/Demo1/TextMesh Pro/Shaders.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: e9f693669af91aa45ad615fc681ed29f
+guid: 0e13b3f941eefe54d929d2ee09958c9e
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
