+++ Assets/Packages/com.studio23.ss2.bettercursormanager/Samples/Demo1/TextMesh Pro/Shaders/TMP_Bitmap-Custom-Atlas.shader.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 48bb5f55d8670e349b6e614913f9d910
+guid: 8f850fb8fe9d23844af576eb59d60720
 ShaderImporter:
   externalObjects: {}
   defaultTextures: []
