+++ Assets/Packages/com.studio23.ss2.bettercursormanager/Samples/Demo1/TextMesh Pro/Shaders/TMP_Bitmap-Mobile.shader.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 1e3b057af24249748ff873be7fafee47
+guid: 3bbb7c90fe0915d44896a8100392265c
 ShaderImporter:
   externalObjects: {}
   defaultTextures: []
