+++ Assets/Packages/com.studio23.ss2.bettercursormanager/Samples/Demo1/TextMesh Pro/Shaders/TMP_Bitmap.shader.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 128e987d567d4e2c824d754223b3f3b0
+guid: cb4f680e77fc8144e84c9c7cafa4e20f
 ShaderImporter:
   externalObjects: {}
   defaultTextures: []
