+++ Assets/Packages/com.studio23.ss2.bettercursormanager/Samples/Demo1/TextMesh Pro/Shaders/TMP_SDF Overlay.shader.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: dd89cf5b9246416f84610a006f916af7
+guid: 8c8de017269e5f64f8790d76e5f211f8
 ShaderImporter:
   externalObjects: {}
   defaultTextures: []
