+++ Assets/Packages/com.studio23.ss2.bettercursormanager/Samples/Demo1/TextMesh Pro/Shaders/TMP_SDF SSD.shader.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 14eb328de4b8eb245bb7cea29e4ac00b
+guid: 83e8472e184cde94a95958b13785a022
 ShaderImporter:
   externalObjects: {}
   defaultTextures: []
