+++ Assets/Packages/com.studio23.ss2.bettercursormanager/Samples/Demo1/TextMesh Pro/Shaders/TMP_SDF-Mobile Masking.shader.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: bc1ede39bf3643ee8e493720e4259791
+guid: 17b7d124b58a08c4e8be7c0e97707307
 ShaderImporter:
   externalObjects: {}
   defaultTextures: []
