+++ Assets/Packages/com.studio23.ss2.bettercursormanager/Samples/Demo1/TextMesh Pro/Shaders/TMP_SDF-Mobile Overlay.shader.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: a02a7d8c237544f1962732b55a9aebf1
+guid: 9202c6c88b5a87a4a9b6fefd817c205a
 ShaderImporter:
   externalObjects: {}
   defaultTextures: []
