+++ Assets/Packages/com.studio23.ss2.bettercursormanager/Samples/Demo1/TextMesh Pro/Shaders/TMP_SDF-Mobile SSD.shader.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: c8d12adcee749c344b8117cf7c7eb912
+guid: 1e760aeb3651a114585c2b66e4a4aec8
 ShaderImporter:
   externalObjects: {}
   defaultTextures: []
