+++ Assets/Packages/com.studio23.ss2.bettercursormanager/Samples/Demo1/TextMesh Pro/Shaders/TMP_SDF-Mobile.shader.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: fe393ace9b354375a9cb14cdbbc28be4
+guid: 765819a5d3f16e8419ba112cd3085756
 ShaderImporter:
   externalObjects: {}
   defaultTextures: []
