+++ Assets/Packages/com.studio23.ss2.bettercursormanager/Samples/Demo1/TextMesh Pro/Shaders/TMP_SDF-Surface-Mobile.shader.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 85187c2149c549c5b33f0cdb02836b17
+guid: 355460465c28d9c4a9cb18e5d994ce6b
 ShaderImporter:
   externalObjects: {}
   defaultTextures: []
