+++ Assets/Packages/com.studio23.ss2.bettercursormanager/Samples/Demo1/TextMesh Pro/Shaders/TMP_SDF-Surface.shader.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: f7ada0af4f174f0694ca6a487b8f543d
+guid: 9d488b9524c16af469306c3aef167529
 ShaderImporter:
   externalObjects: {}
   defaultTextures: []
