+++ Assets/Packages/com.studio23.ss2.bettercursormanager/Samples/Demo1/TextMesh Pro/Shaders/TMP_SDF.shader.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 68e6db2ebdc24f95958faec2be5558d6
+guid: 3cf8f7ce2a02d7c4980646fce3b6ad19
 ShaderImporter:
   externalObjects: {}
   defaultTextures: []
