+++ Assets/Packages/com.studio23.ss2.bettercursormanager/Samples/Demo1/TextMesh Pro/Shaders/TMP_Sprite.shader.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: cf81c85f95fe47e1a27f6ae460cf182c
+guid: 1d0b2c364713d85428a4dd8dcde041f1
 ShaderImporter:
   externalObjects: {}
   defaultTextures: []
