+++ Assets/Packages/com.studio23.ss2.bettercursormanager/Samples/Demo1/TextMesh Pro/Shaders/TMPro.cginc.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 407bc68d299748449bbf7f48ee690f8d
+guid: 526db43ffcbfafa4b8233db7ff81a667
 ShaderImporter:
   externalObjects: {}
   defaultTextures: []
