+++ Assets/Packages/com.studio23.ss2.bettercursormanager/Samples/Demo1/TextMesh Pro/Shaders/TMPro_Mobile.cginc.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: c334973cef89a9840b0b0c507e0377ab
+guid: 5dd52ce133b4ecb44945b7e045f207b3
 ShaderImporter:
   externalObjects: {}
   defaultTextures: []
