+++ Assets/Packages/com.studio23.ss2.bettercursormanager/Samples/Demo1/TextMesh Pro/Shaders/TMPro_Properties.cginc.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 3997e2241185407d80309a82f9148466
+guid: 5b4000628f0733847af352375eccbf6b
 ShaderImporter:
   externalObjects: {}
   defaultTextures: []
