+++ Assets/Packages/com.studio23.ss2.bettercursormanager/Samples/Demo1/TextMesh Pro/Shaders/TMPro_Surface.cginc.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: d930090c0cd643c7b55f19a38538c162
+guid: dfad41c5f013621499db24a69a4faa42
 ShaderImporter:
   externalObjects: {}
   defaultTextures: []
