+++ Assets/Packages/com.studio23.ss2.bettercursormanager/Samples/Demo1/TextMesh Pro/Sprites.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: d0603b6d5186471b96c778c3949c7ce2
+guid: c18fcc1d1b255a6479107ddb933dbab3
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
