+++ Assets/Packages/com.studio23.ss2.bettercursormanager/Samples/Demo1/TextMesh Pro/Sprites/EmojiOne Attribution.txt.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 381dcb09d5029d14897e55f98031fca5
+guid: 8538b0fff75c815439c3d676942dc5f8
 TextScriptImporter:
   externalObjects: {}
   userData:
