+++ Assets/Packages/com.studio23.ss2.bettercursormanager/Samples/Demo1/TextMesh Pro/Sprites/EmojiOne.json.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 8f05276190cf498a8153f6cbe761d4e6
+guid: 738a33e138f1d464a9946727647de34a
 timeCreated: 1480316860
 licenseType: Pro
 TextScriptImporter:
