+++ Assets/Packages/com.studio23.ss2.bettercursormanager/Samples/Demo1/TextMesh Pro/Sprites/EmojiOne.png.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: dffef66376be4fa480fb02b19edbe903
+guid: 32f60e887b2d08f4e8d412f124feddd9
 TextureImporter:
   fileIDToRecycleName:
     21300000: EmojiOne_0
